+++ src/clojupyter/kernel/logo.clj
@@ -3,29 +3,30 @@
    [clojure.edn			:as edn]
    [clojure.java.io		:as io]
    ,,
-   [clojupyter.display		:as dis])
+   [clojupyter.display		:as dis]
+)
   (:import
    [javax.imageio ImageIO]))
 
-(def ^:private logo-filename	"(clojupyter) 350x80.png")
-(def ^:private license-filename	"LICENSE.TXT")
+(def ^:private LOGO-PATH	"clojupyter/assets/logo-350x80.png")
+(def ^:private LICENSE-PATH	"clojupyter/assets/license.txt")
 
 (defn- logo-dimensions
   [filename]
-  (let [[_ x y] (re-find #"(\d+)x(\d+).png$" filename)]
+  (when-let [[_ x y] (re-find #"(\d+)x(\d+).png$" filename)]
     (mapv edn/read-string [x y])))
 
 (defn logo-resource
   "Returns a clojupyter logo as a `java.io.resource`."
   []
-  (or (io/resource logo-filename)
-      (->> logo-filename (str "./resources/") io/file)))
+  (or (io/resource LOGO-PATH)
+      (->> LOGO-PATH (str "./resources/") io/file)))
 
 (defn license-resource
   "Returns the clojupyter license as a `java.io.resource`."
   []
-  (or (io/resource license-filename)
-      (->> license-filename (str "./resources/") io/file)))
+  (or (io/resource LICENSE-PATH)
+      (->> LICENSE-PATH (str "./resources/") io/file)))
 
 (defn logo-image
   "Returns a clojupyter logo as a `java.awt.image.BufferedImage`."